// packages/contract-helpers/src/V3-uiStakeDataProvider-contract-copy/types.ts
export type StakedTokenData = {
  stakedTokenTotalSupply: BigNumber;
  stakedTokenTotalRedeemableAmount: BigNumber;
  stakeCooldownSeconds: BigNumber;
  stakeUnstakeWindow: BigNumber;
  rewardTokenPriceEth: BigNumber;
  distributionEnd: BigNumber;
  distributionPerSecond: BigNumber;
  stakedTokenPriceEth: BigNumber;
  stakeApy: BigNumber;
  inPostSlashingPeriod: boolean;
};

export type HumanizedStakeResponse = {
  parsedStakedData: StakedTokenData[];
  parsedPrices: string[];
  ethPriceUsd: string;
};